Added quickSort into CurveHelper.
