+++ src/GShark/Core/CurveHelpers.cs
@@ -1,4 +1,5 @@
-﻿using GShark.ExtendedMethods;
+﻿using System;
+using GShark.ExtendedMethods;
 using GShark.Geometry;
 using GShark.Operation;
 using System.Collections.Generic;
@@ -45,6 +46,100 @@ namespace GShark.Core
                     : curve).ToList();
 
             return curves;
+        }
+
+        /// <summary>
+        /// A quickSort algorithm used to order the curves in a sequential manner based on the distance between the first and last points of each curves.<br/>
+        /// https://github.com/mcneel/opennurbs/blob/c20e599d1ff8f08a55d3dddf5b39e37e8b5cac06/opennurbs_curve.cpp#L3600 <br/>
+        /// https://exceptionnotfound.net/quick-sort-csharp-the-sorting-algorithm-family-reunion/
+        /// </summary>
+        /// <param name="curves">The sets of curve to sort.</param>
+        /// <returns>The set of curves sorted.</returns>
+        public static List<NurbsCurve> QuickSortCurve(IList<NurbsCurve> curves)
+        {
+            if (curves == null || curves.Count == 0)
+            {
+                throw new Exception("The set of curves is empty.");
+            }
+
+            if (curves.Count == 1)
+            {
+                return curves.ToList();
+            }
+
+            // creates the sets of data required.
+            List<Point3[]> lines = curves.Select(c => new[] { c.StartPoint, c.EndPoint }).ToList();
+            int[] indexes = new int[lines.Count];
+            bool[] revers = new bool[lines.Count];
+            for (int t = 0; t < lines.Count; t++)
+            {
+                revers[t] = false;
+                indexes[t] = t;
+            }
+
+            // sort lines
+            for (int ni = 1; ni < lines.Count; ni++)
+            {
+                int endI, endEnd, i;
+                var startI = endI = ni;
+                var startEnd = endEnd = 0;
+                var startPoint = (revers[0]) ? lines[indexes[0]][1] : lines[indexes[0]][0];
+                var endPoint = (revers[ni - 1]) ? lines[indexes[ni - 1]][0] : lines[indexes[ni - 1]][1];
+                var startDistance = startPoint.DistanceTo(lines[indexes[startI]][0]);
+                var endDistance = endPoint.DistanceTo(lines[indexes[endI]][0]);
+
+                for (i = ni; i < lines.Count; i++)
+                {
+                    Point3 testingPoint = lines[indexes[i]][0];
+                    for (int end = 0; end < 2; end++)
+                    {
+                        double testingDistance = startPoint.DistanceTo(testingPoint);
+                        if (testingDistance < startDistance)
+                        {
+                            startI = i;
+                            startEnd = end;
+                            startDistance = testingDistance;
+                        }
+
+                        testingDistance = endPoint.DistanceTo(testingPoint);
+                        if (testingDistance < endDistance)
+                        {
+                            endI = i;
+                            endEnd = end;
+                            endDistance = testingDistance;
+                        }
+
+                        testingPoint = lines[indexes[i]][1];
+                    }
+                }
+
+                if (startDistance < endDistance)
+                {
+                    // N[index[startI]] will be first in list.
+                    i = indexes[ni];
+                    indexes[ni] = indexes[startI];
+                    indexes[startI] = i;
+                    startI = indexes[ni];
+                    for (i = ni; i > 0; i--)
+                    {
+                        indexes[i] = indexes[i - 1];
+                        revers[i] = revers[i - 1];
+                    }
+                    indexes[0] = startI;
+                    revers[0] = (startEnd != 1);
+                }
+                else
+                {
+                    // N[index[endI]] will be next in the list.
+                    i = indexes[ni];
+                    indexes[ni] = indexes[endI];
+                    indexes[endI] = i;
+                    revers[ni] = (endEnd == 1);
+                }
+            }
+
+            List<NurbsCurve> sortedCurves = indexes.Select(i => (revers[i]) ? curves[i].Reverse() : curves[i]).ToList();
+            return sortedCurves;
         }
     }
 }
